# Use fast_pow for faster srgb/linear conversions (#1908)