#1908 
Test results with [http://entropymine.com/imageworsener/slowpow/](http://entropymine.com/imageworsener/slowpow/):

**Original pow:**

```
$ time; ./powtest 100000000 1.0000000000000020 1.5000050000000000 ; time
shell  0.04s user 0.03s system 0% cpu 24.356 total
children  0.06s user 0.01s system 0% cpu 24.356 total
(count=100000000) 1.0000000000000020^1.5000050000000000 = 1.0000000000000031
shell  0.04s user 0.03s system 0% cpu 30.206 total
children  5.86s user 0.03s system 19% cpu 30.206 total
```
- Test 1: ~6 seconds
- Test 2 and 3: way too long

**Appleseed's fast pow:**
```
oktomus at pangolin in ~ 
$ time; ./fastpowtest 100000000 1.0000000000000020 1.5000050000000000 ; time
shell  0.05s user 0.03s system 0% cpu 1:18.20 total
children  5.86s user 0.03s system 7% cpu 1:18.20 total
(count=100000000) 1.0000000000000020^1.5000050000000000 = 1.0000152587890625
shell  0.05s user 0.03s system 0% cpu 1:20.92 total
children  8.56s user 0.03s system 10% cpu 1:20.92 total

oktomus at pangolin in ~ 
$ time; ./fastpowtest 100000000 1.0000000000000020 1.5000005000000000  ; time
shell  0.06s user 0.04s system 0% cpu 1:32.03 total
children  8.56s user 0.03s system 9% cpu 1:32.03 total
(count=100000000) 1.0000000000000020^1.5000005000000001 = 1.0000152587890625
shell  0.06s user 0.04s system 0% cpu 1:34.74 total
children  11.25s user 0.03s system 11% cpu 1:34.74 total

oktomus at pangolin in ~ 
$ time; ./fastpowtest 100000000 1.0000000000000020 1.5000000000000000   ; time
shell  0.06s user 0.04s system 0% cpu 1:40.89 total
children  11.25s user 0.03s system 11% cpu 1:40.89 total
(count=100000000) 1.0000000000000020^1.5000000000000000 = 1.0000152587890625
shell  0.06s user 0.04s system 0% cpu 1:43.60 total
children  13.95s user 0.04s system 13% cpu 1:43.60 total
```
- Test 1: ~2 seconds
- Test 2: ~2 seconds
- Test 3: ~3 seconds

**OIIO's safe fast pow:**
```
$ time; ./oiiofastpowtest 100000000 1.0000000000000020 1.5000050000000000    ; time
shell  0.07s user 0.05s system 0% cpu 3:31.59 total
children  13.95s user 0.04s system 6% cpu 3:31.59 total
(count=100000000) 1.0000000000000020^1.5000050000000000 = 1.0000000000000000
shell  0.07s user 0.05s system 0% cpu 3:50.09 total
children  32.41s user 0.04s system 14% cpu 3:50.09 total
```
- Test 1: ~19 seconds
- Test 2 and 3: way to long

Tested also with the cornell box scene, saving around 4 seconds and the result is the same.