@dictoon 2018-03-25
So I'm finally having a closer look at this PR :)

We don't want to change `linear_rgb_to_srgb()`. You can see it as the "ground truth": we don't care too much if it's slow, what we want is it to be as precise as possible.

When performance matters, `fast_linear_rgb_to_srgb()` should be used instead.